Use non-overlapped sockets on Windows by default
OCaml currently exposes only overlapped sockets. Overlapped sockets
are created with the `socket` function or with `WSASocket` by using
the `WSA_FLAG_OVERLAPPED` flag [1].

They have one caveat: non-overlapped sockets cannot be used for IO
redirection. See below for a C snippet showing that it won't work.
I haven't found documentation for this behaviour except forum posts
[2]. The use case is spawning a process using `CreateProcess` and
substituting a standard handle with a socket (e.g., a Unix domain
socket).

It isn't possible to change the "overlapped-ness" of a socket after
creation, so the possibility to chose between overlapped and
non-overlapped must be offered when creating a socket. Two functions
that create sockets are exposed in OCaml: `socket` and `socketpair`.

This proposal adds type `socket_flags` for platform-specific
socket-related flags, and an optional parameter `?flags: socket_flags
list`. The intend is that this parameter allows for future extension.

Now should OCaml offer overlapped or non-overlapped sockets by
default? I chose to align with the WSA api and go with non-overlapped
sockets, but that is a breaking change and may raise another whole lot
of issues.

[1]: https://docs.microsoft.com/en-us/windows/win32/winsock/overlapped-i-o-and-event-objects-2
[2]: https://stackoverflow.com/q/4993119/3566655