Add a -source-tree option to ocamldep
If ocamldep is invoked without `-all`, then the output doesn't refer
to source files and there is no need for `-source-tree`.

If `-all` is used, though, then rules output by ocamldep will refer to
source files and, currently, they will contain the paths to the source
files relative to where ocamldep is run from.

Defining `-source-tree` allows to run the build from a different
directory than of which ocamldep was run.