Annotate errors using GitHub Actions markup in ocamltest
It can be a bit tedious to locate a failing test in the whole
testsuite, scrolling endlessly in the GitHub Actions interface. It's
possible, using a simple markup language documented at [1], to tell
GitHub to highlight a location in the logs. Detecting the CI is
achieved using two environment variables, `CI`, and
`GITHUB_ACTIONS`. Making full use of the markup language (i.e.,
linking to the source file where the test fails, showing advanced
messages in the GHA UI) is currenlty not possible because these
information are not extracted from the test run.

[1]: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions