Properly detect mingw-w64 as a C compiler vendor
Contrary to the comment in `configure.ac`, it is possible to detect
the use MinGW-w64 as a C compiler with the `__MINGW32__` macro, which
we're already using throughout the compiler. A header is needed to
retrieve MinGW-w64 version (distinct from GCC). [Preprocessor
output][1] needs to be trimmed.

MinGW-w64 can also be [coupled to LLVM instead of GCC][2]; to be able
to differentiate the backend compiler and tell its version, the "real"
C compiler vendor information is also included.

As an example, some possible values:
- mingw32-10-0-gcc-11-3,
- mingw32-12-0-clang-17-0.

[1]: https://gcc.gnu.org/onlinedocs/cpp/Preprocessor-Output.html
[2]: https://www.mingw-w64.org/downloads/

(cherry picked from commit 65466857701cf695eac8f9d019aab0e48853c3ce)