appveyor: evict the configure cache sometimes
The configure cache might be stale if the configure script has
changed: the script responsible for a cache entry might have been
modified. Try to detect whether configure has changed, and if so,
evict the cache.

On a pull request because AppVeyor exposes the base of the PR, so it's
simple to check for a change in the revision range, but on a push it
doesn't expose the commit ID of the previous state of the ref. In that
case, it's safer to always evict the cache.