Forbid non-standard C flexible array member
Use -fstrict-flex-arrays=3 (supported since GCC 13, clang 17) when
compiling C. Useful for instrumented bounds-checking, and makes the
code clearer.

LEVEL=3 is the strictest level, only when the trailing array is
declared as a flexible array member per C99 standard onwards ([]), it
is treated as a flexible array member.
When LEVEL is 2, the trailing array is treated as a flexible array
member when it is declared as either [], or [0].

On macOS, clang 15 is the default at the time of writing, and only
supports LEVEL=2.

It's not possible to use this flag with a level > 1 on Windows with
clang-cl or MinGW-w64, as the WinAPI often use one-sized flexible
array members for historical reasons (see e.g., REPARSE_DATA_BUFFER).

[Why do some structures end with an array of size 1?](https://devblogs.microsoft.com/oldnewthing/?p=38043)