Support spin-wait sleeps with sub-microsecond precision on Windows
By using a high-precision timer, available since Windows 10
RS4 (1803), we can attain sleeps with a 100 nanosecond precision. The
only caveat is that a handle is required for a thread to sleep. As the
handle can be reused, we store it in the domain state, for caching, in
order not to pay the cost of creating and closing it when calling the
new sleep function.

Co-authored-by: Samuel Hym <samuel.hym@rustyne.lautre.net>