On Windows, force the C preproc to treat input files as C
clang defines:

> -E   Only run the preprocessor
> -P   Disable linemarker output in -E mode

msvc defines:

> /E   preprocess to stdout
> /P   preprocess to file
> /EP  preprocess to stdout, no #line

Thus, msvc's /EP is equivalent to clang's -E -P.

However, clang-cl gets confused if it can't infer that the input file
is C, transforms msvc's /EP to msvc's /E /P and discards /E, writing
the output of the preprocessor with linemarkers to a file.

Circumvent the problem by telling the C preprocessor it's processing a C
file. This also works with .ml.c and .mli.c files.