Define unreachable() (before C23) to mark unreachable code
Use the unreachable annotation to optimize impossible code branches
away. If there's no sensible definition available, fall back to
abort. The macro is part of stddef.h as of C23.

https://en.cppreference.com/w/c/program/unreachable

cppreference currently suggests a possible implementation:

    [[noreturn]] inline void unreachable_impl() {}
    #define unreachable() (unreachable_impl())

but we cannot use a noreturn void function with an empty body, as the
compiler may warn that it has an implicit return statement.

For the debug mode, we could want to trap the program to catch bugs
and prevent further execution, instead of optimizing unreachable
branches. A possible implementation would be:

    #if DEBUG
      #undef unreachable
      #if __has_builtin(__builtin_trap)
        #define unreachable() (__builtin_trap())
      #elif defined(_MSC_VER)
        #include <intrin.h>
        CAMLnoret Caml_inline void caml_fastfail(unsigned int);
        void caml_fastfail(unsigned int i) { __fastfail(i); }
        #define unreachable() (caml_fastfail(7 /* FAST_FAIL_FATAL_APP_EXIT */))
      #else
        #define unreachable() (CAMLassert(0))
      #endif
    #endif

MSVC currently has a bug with unreachable code analysis and its
__fastfail intrinsic, thus the noreturn wrapper function.

https://developercommunity.visualstudio.com/t/C-Code-Analysis-should-understand-that/10665570