gha: use lld-link as a linker with clang-cl
The flag -fuse-ld=lld-link cannot be part of LDFLAGS, because LDFLAGS
are given to the _compiler driver_ after /link, where flags are given
as-is to the linker. The linker has already been selected at this
point. It cannot be part of CFLAGS either, because CFLAGS shouldn't be
passed when invoking the linker (they're for C compilation). Thus, it
has to be part of CC, changing the _compiler driver_ invocation.