Create symlinks on Windows when available
Previously Windows unconditionally uses `cp`, doubling the size
required for the OCaml binaries. `configure` now determines if `ln`
creates native symlinks and only uses `cp` if that fails. Users of the
compiler are simply required to enable Developer Mode (or build OCaml
using an elevated shell) to benefit from symlinks.

https://cygwin.com/cygwin-ug-net/using.html#pathnames-symlinks
https://www.msys2.org/docs/symlinks/

Harden the implementation by setting the `CYGWIN` and `MSYS`
environment variables in stone in the Makefile.