Support choosing an alternative linker through the C compiler
GCC, clang, and clang-cl accept the `-fuse-ld=<ld>` option to call a
specific linker. This can be passed through the LDFLAGS variable when
the linker is invoked through the compiler driver.

When calling the linker through the cl.exe or clang-cl.exe compiler
drivers, linker flags (LDFLAGS) have to be passed after the `/link`
flag. Unfortunately, the `-fuse-ld=<ld>` option of clang-cl has to be
passed to clang-cl *before* `/link`, as this option is needed by the
compiler driver, rather than the linker. This requires a bit of hacks.